MAJOR RE-STRUCTURE
Modules and everything working with a lot of other program flow changes
