+++ pages/friend_search.py
@@ -1,0 +1,13 @@
+from common_utils.utils import clearScreen
+
+# user selected to find someone that you know
+def printFriendSearchScreen():
+    clearScreen()
+    print("*** Find A Friend ***")
+    print("under construction, input anything to return")
+    userInput = input()
+    # printMainMenu()  # Delete me
+    return # this returns to caller so this function doesn't need to import printMainMenu()
+    # with printMainMenu(), there is a circular import between main_menu.py and friend_search.py
+
+
