+++ pages/job_search.py
@@ -1,0 +1,14 @@
+from common_utils.utils import clearScreen
+# from pages.main_menu import printMainMenu
+
+# user selected to do a job search
+def printJobSearchScreen():
+    clearScreen()
+    print("*** Job Search ***")
+    print("under construction, input anything to return")
+    userInput = input()
+    # printMainMenu()  # Delete me
+    # causing circular import error
+    return
+
+
